refactor: 기존 open api response dto rename 및 위치 이동
