+++ src/main/java/until/the/eternity/auction/domain/component/AuctionHistoryFetcher.java
@@ -12,8 +12,8 @@ import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.reactive.function.client.WebClientResponseException;
 import reactor.core.publisher.Mono;
 import reactor.util.retry.Retry;
-import until.the.eternity.auction.domain.dto.AuctionHistoryDto;
-import until.the.eternity.auction.domain.dto.AuctionHistoryResponse;
+import until.the.eternity.auction.domain.dto.external.OpenApiAuctionHistoryResponse;
+import until.the.eternity.auction.domain.dto.external.OpenApiAuctionHistoryListResponse;
 import until.the.eternity.auction.domain.repository.AuctionHistoryRepository;
 import until.the.eternity.common.enums.ItemCategory;
 
@@ -28,15 +28,15 @@ public class AuctionHistoryFetcher {
     @Value("${openapi.nexon.api-key}")
     private String nexonApiKey;
 
-    public List<AuctionHistoryDto> fetch(ItemCategory category) {
-        List<AuctionHistoryDto> result = new ArrayList<>();
+    public List<OpenApiAuctionHistoryResponse> fetch(ItemCategory category) {
+        List<OpenApiAuctionHistoryResponse> result = new ArrayList<>();
         String cursor = null;
 
         do {
-            AuctionHistoryResponse response = fetchFromApi(category, cursor);
+            OpenApiAuctionHistoryListResponse response = fetchFromApi(category, cursor);
             if (response == null || response.getAuction_history() == null) break;
 
-            List<AuctionHistoryDto> currentBatch = response.getAuction_history();
+            List<OpenApiAuctionHistoryResponse> currentBatch = response.getAuction_history();
             if (containsExistingIds(currentBatch)) break;
 
             result.addAll(currentBatch);
@@ -46,7 +46,7 @@ public class AuctionHistoryFetcher {
         return result;
     }
 
-    private AuctionHistoryResponse fetchFromApi(ItemCategory category, String cursor) {
+    private OpenApiAuctionHistoryListResponse fetchFromApi(ItemCategory category, String cursor) {
         try {
             return webClient
                     .get()
@@ -64,7 +64,7 @@ public class AuctionHistoryFetcher {
                     .header("x-nxopen-api-key", nexonApiKey)
                     .header("accept", "application/json")
                     .retrieve()
-                    .bodyToMono(AuctionHistoryResponse.class)
+                    .bodyToMono(OpenApiAuctionHistoryListResponse.class)
                     .retryWhen(
                             Retry.backoff(3, Duration.ofSeconds(2)) // 🔁 최대 3번 재시도, 2초 간격 (지수 백오프)
                                     .filter(this::isRetryableException)) // 재시도 조건
@@ -89,9 +89,9 @@ public class AuctionHistoryFetcher {
                 && ((WebClientResponseException) throwable).getStatusCode().is5xxServerError();
     }
 
-    private boolean containsExistingIds(List<AuctionHistoryDto> dtos) {
+    private boolean containsExistingIds(List<OpenApiAuctionHistoryResponse> dtos) {
         List<String> ids =
-                dtos.stream().map(AuctionHistoryDto::getAuctionBuyId).collect(Collectors.toList());
+                dtos.stream().map(OpenApiAuctionHistoryResponse::getAuctionBuyId).collect(Collectors.toList());
         return auctionHistoryRepository.existsByAuctionBuyIdIn(ids);
     }
 }
