+++ src/main/java/until/the/eternity/auction/domain/dto/external/OpenApiAuctionHistoryListResponse.java
@@ -1,13 +1,13 @@
-package until.the.eternity.auction.domain.dto;
+package until.the.eternity.auction.domain.dto.external;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
 import lombok.Data;
 
 @Data
-public class AuctionHistoryResponse {
+public class OpenApiAuctionHistoryListResponse {
     @JsonProperty("auction_history")
-    private List<AuctionHistoryDto> auction_history;
+    private List<OpenApiAuctionHistoryResponse> auction_history;
 
     @JsonProperty("next_cursor")
     private String next_cursor;
