+++ src/main/java/until/the/eternity/auction/domain/repository/AuctionHistoryRepositoryCustom.java
@@ -2,7 +2,7 @@ package until.the.eternity.auction.domain.repository;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import until.the.eternity.auction.domain.dto.AuctionHistorySearchCondition;
+import until.the.eternity.auction.domain.dto.internal.request.AuctionHistorySearchCondition;
 import until.the.eternity.auction.domain.model.AuctionHistory;
 
 public interface AuctionHistoryRepositoryCustom {
